[PLAT-8015] Fix all string format issues in platform code
Summary: Wrong format specifiers corrected + some minor cleanups.

Test Plan: Minor. UT + Itests

Reviewers: cvankayalapati, cwang

Reviewed By: cwang

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D23898

+++ managed/src/main/java/com/yugabyte/yw/common/AccessManager.java
@@ -297,13 +297,15 @@ public class AccessManager extends DevopsBase {
       log.error(ioe.getMessage(), ioe);
       throw new RuntimeException("Could not create AccessKey", ioe);
     } finally {
-      try {
-        File tmpKeyFile = new File(tempFile.toString());
-        if (tmpKeyFile.exists()) {
-          Files.delete(tempFile);
+      if (tempFile != null) {
+        try {
+          File tmpKeyFile = new File(tempFile.toString());
+          if (tmpKeyFile.exists()) {
+            Files.delete(tempFile);
+          }
+        } catch (IOException e) {
+          log.error(e.getMessage(), e);
         }
-      } catch (IOException e) {
-        log.error(e.getMessage(), e);
       }
     }
